changing scripts

+++ setup.sh
@@ -7,29 +7,31 @@ fi
 
 export ORG="acend"
 export APP="theia"
-export USERS=2
+export STUDENTS=2
 
 build() {
-    set -e
-    docker build -t $ORG/$APP .
-    docker push $ORG/$APP
-    docker image prune --force
-    set +e
+    if [ -n "$(which docker)" ]; then
+        docker build -t $ORG/$APP .
+        docker push $ORG/$APP
+        docker image prune --force
+    elif [ -n "$(which buildah)" ]; then
+        sudo buildah build -t $ORG/$APP .
+        sudo buildah push $ORG/$APP
+        sudo buildah image prune --force
+    fi
 }
 
 deploy() {
-    for i in $(seq 1 $USERS); do
-        export USER=user$i
-        # kubectl -n $USER scale deploy $USER-shell-deploy --replicas=0
+    for i in $(seq 1 $STUDENTS); do
+        export STUDENT=student$i
         cat deliver/workspace.yaml | envsubst | kubectl apply -f -
     done
-    # kubectl -n train rollout status deployment shell-deploy
 }
 
 destroy() {
     for i in $(seq 1 $USERS); do
-        export USER=user$i
-        kubectl delete ns $USER
+        export STUDENT=student$i
+        kubectl delete ns $STUDENT
     done
 }
